requirements.txt
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.7
filelock==3.9.0
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
platformdirs==3.1.0
protobuf==3.20.3
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.20.0
Werkzeug==2.2.3
zipp==3.15.0
